requirements.txt
brotlipy
gdown
omegaconf
pip-chill
pyopenssl
pysocks
pytest
torchaudio
torchvision
transformers
emoji
nltk
gradio==3.8.0
flask
flask_healthz
rdflib
rdflib-jsonld
pysolr==3.8.1
numpy==1.17.2
segtok==1.5.7
colorama==0.4.3
termcolor==1.1.0
pytorch_transformers
requests-cache
